Issue 6 merged pravega writer and exactly once writer code, used checkpoint to decide on transaction api usage
